refactor: pull data model functions out of soilIdSlice, into new file
